+++ src/main/java/com/portfolio/www/index/IndexController.java
@@ -15,42 +15,12 @@ import com.portfolio.www.service.IndexService;
 @Controller
 public class IndexController {
 	
-	@Autowired
-	IndexService indexService;
-	
-	// 인덱스페이지 개별 메뉴 추천
-	@ResponseBody
-	@RequestMapping(value = "/recomMenu.do", produces = "text/plain; charset=UTF-8") // 한글인식 안되는 문제 해결
-	public String recomMenu(@RequestParam String category) {
-		
-		String menu = ""; 
-		
-		if(category.equals("아무거나")) {
-			menu = indexService.allMenu();
-		} else {
-			menu = indexService.getMenu(category);
-		}
-		
-		return menu;
-	}
-	
-	
 	// 인덱스페이지
 	@RequestMapping("/index.do")
 	public ModelAndView index(@RequestParam HashMap<String, String> params) {
 		ModelAndView mv = new ModelAndView();
-		
 		System.out.println("================ IndexContoller 진입 ================");
-
 		mv.setViewName("index");
-		
-		// 모든 메뉴 카테고리 가져오기
-		List<String> category = indexService.menuCategory();
-		mv.addObject("category",category);
-		
-		// 모든 메뉴 가져오기
-		String menu = indexService.allMenu();
-		mv.addObject("menu",menu);
 		return mv;
 	}
 }
